+++ api/operator/v1beta1/vmcluster_types_test.go
@@ -8,6 +8,7 @@ import (
 
 func TestVMBackup_SnapshotDeletePathWithFlags(t *testing.T) {
 	type opts struct {
+		host      string
 		port      string
 		extraArgs map[string]string
 		want      string
@@ -15,27 +16,38 @@ func TestVMBackup_SnapshotDeletePathWithFlags(t *testing.T) {
 	f := func(o opts) {
 		t.Helper()
 		cr := VMBackup{}
-		if got := cr.SnapshotDeletePathWithFlags(o.port, o.extraArgs); got != o.want {
+		if got := cr.SnapshotDeletePathWithFlags(o.host, o.port, o.extraArgs); got != o.want {
 			t.Errorf("SnapshotDeletePathWithFlags() = %v, want %v", got, o.want)
 		}
 	}
 
 	// default delete path
 	f(opts{
+		host: "localhost",
 		port: "8428",
 		want: "http://localhost:8428/snapshot/delete",
 	})
 
 	// delete path with prefix
 	f(opts{
+		host:      "127.0.0.1",
 		port:      "8428",
 		extraArgs: map[string]string{vmPathPrefixFlagName: "/pref-1", "other-flag": "other-value"},
-		want:      "http://localhost:8428/pref-1/snapshot/delete",
+		want:      "http://127.0.0.1:8428/pref-1/snapshot/delete",
+	})
+
+	// delete path with auth key
+	f(opts{
+		host:      "127.0.0.1",
+		port:      "8428",
+		extraArgs: map[string]string{vmPathPrefixFlagName: "/pref-1", "other-flag": "other-value", snapshotAuthKey: "test"},
+		want:      "http://127.0.0.1:8428/pref-1/snapshot/delete?authKey=test",
 	})
 }
 
 func TestVMBackup_SnapshotCreatePathWithFlags(t *testing.T) {
 	type opts struct {
+		host      string
 		port      string
 		extraArgs map[string]string
 		want      string
@@ -43,27 +55,30 @@ func TestVMBackup_SnapshotCreatePathWithFlags(t *testing.T) {
 	f := func(o opts) {
 		t.Helper()
 		cr := VMBackup{}
-		got := cr.SnapshotCreatePathWithFlags(o.port, o.extraArgs)
+		got := cr.SnapshotCreatePathWithFlags(o.host, o.port, o.extraArgs)
 		assert.Equal(t, o.want, got)
 	}
 
 	// base ok
 	f(opts{
+		host: "localhost",
 		port: "8429",
 		want: "http://localhost:8429/snapshot/create",
 	})
 
 	// with prefix
 	f(opts{
+		host: "127.0.0.1",
 		port: "8429",
 		extraArgs: map[string]string{
 			"http.pathPrefix": "/prefix/custom",
 		},
-		want: "http://localhost:8429/prefix/custom/snapshot/create",
+		want: "http://127.0.0.1:8429/prefix/custom/snapshot/create",
 	})
 
 	// with prefix and auth key
 	f(opts{
+		host: "localhost",
 		port: "8429",
 		extraArgs: map[string]string{
 			"http.pathPrefix": "/prefix/custom",